requirements.txt
aiohttp==3.13.3
av==15.1.0
cmake==4.2.1
cuda-bindings==12.9.4
datasets==4.1.1
deepdiff==8.6.1
diffusers==0.18.2
dm_control==1.0.20
draccus==0.10.0
einops==0.8.2
fvcore==0.1.5.post20221221
gdown==5.2.1
gym==0.26.2
gymnasium==1.2.3
hf_transfer==0.1.9
hf-xet==1.2.0
httpx==0.28.1
importlib_resources==6.5.2
inquirerpy==0.3.4
jsonlines==4.0.0
Mako==1.3.10
ninja==1.13.0
nvidia-cufile-cu12==1.13.1.3
nvidia-cusparselt-cu12==0.7.1
nvidia-nvshmem-cu12==3.4.5
opencv-python-headless==4.12.0.88
osfclient==0.0.5
peft==0.10.0
pip==26.0.1
pygame==2.6.1
pylint==4.0.4
pymunk==6.5.1
pyserial==3.5
PySocks==1.7.1
pytorch3d==0.7.8
quadprog==0.1.13
rerun-sdk==0.26.2
robomimic==0.3.0
robosuite==1.5.2
scikit-image==0.25.2
scikit-video==1.1.11
shapely==2.1.2
shellingham==1.5.4
terminaltables==3.1.10
torchaudio==2.2.2+cu121
torchcfm==1.0.7
torchcodec==0.5
typer-slim==0.21.1
wandb==0.24.2
wheel==0.46.3
zarr==2.18.3
